Adding Bengali support
- added proper OpenSSL libs for updater
- fixed some compilation warnings
- added support for Bengali
- removed command line when application starts.

+++ projects/sargam/utils/CommandStack.cpp
@@ -40,7 +40,7 @@ void CommandStack::add(Command* ipC)
   
   //on ajoute et execute la commande insérée
   mCommands.push_back(ipC);
-  mIndex = mCommands.size() - 1;
+  mIndex = (int)mCommands.size() - 1;
 }
 
 //------------------------------------------------------------------------------
